requirements.txt
flask>=2.0.0
pillow>=8.0.0
requests>=2.25.0
